refactor(ws): fix gateway cross-layer imports
- Move relayDmToBot from handlers/dm.handler.ts to lib/dm-relay.ts
- Both handler and gateway now import from the shared lib layer
- Remove direct handler imports from gateways
- Maintain proper Handler → Service → DAO architecture

Found during architecture audit review.
